HFIO

Description

This module contains helper functions for encoding and decoding to file. Actual writing to file only happens in the application side of this library This module was created for the Utrecht University of Applied Sciences AI course 'Declarative Programming'

There are a couple of references to 'delimiter1' and 'delimiter2'. These are defined as follows:

delimiter1 = "||"

delimiter2 = "#|"

Documentation

keyFromString

Arguments

:: String
String, example: OIOI

-> [Binary]
List of Binary, example: [O,I,O,I]

Takes a string containing 'I' or 'O's, returns list of Binary.

reshapeCodes

Arguments

:: [(Char, [Binary])]
Huffman Map in tuple form, example: [('c', [I,O,I])]

-> [(Char, String)]
Huffman Map in Tuple form, example: [('c', "IOI")]

Reshapes Huffman Map

formatCodes

Arguments

:: [(Char, String)]
Huffman Map in tuple form, example: [('c', IOI), ('d', IIO)]

-> String
Stringified Huffman map for printing to file, example: "c||IOI#|/d||IIO"

Takes a list of Huffman Key-Value pairs, turns into String for printing to file. Uses delimiter1 and delimiter2.

listifyCodes

Arguments

:: String
Stringified Huffman map, example: "c||IOI#|/d||IIO"

-> [[String]]
Two dimensional list containing String-code pairs, example: [["c", "IOI"], ["d", "IIO"]]

Inverse of formatCodes using lists instead of tuples

readCodes

Arguments

:: [[String]]
Two dimensional list containg String-code pairs, example [["c", "IOI"], ["d", "IIO"]]

-> [([Binary], Char)]
Huffman Map in tuple form, example: [([I,O,I], 'c'), ([I,I,I], 'd')]

Takes two-dimensional list, returns tuple pair (Huffman Map)

readCodeString

Arguments

:: String
Stringified Huffman map, example: "c||IOI#|/d||IIO"

-> [([Binary], Char)]
Huffman Map in tuple form, example: [([I,O,I], 'c'), ([I,I,I], 'd')]

Composition of readCodes and listifyCodes