Lib

Documentation

countValues

Arguments

:: String
The text to be encoded

-> [(Char, Int)]
(Unique character, frequency of said character)

The countValues function counts each unique character in a string. Adapted from https://stackoverflow.com/a/3711137.

createNodes

Arguments

:: [(Char, Int)]
Unique character is text, frequency of character in text

-> [HuffmanTree]
List of Node

The createNodes function creates a Node for each tuple. The frequency of a character equals its Weight.

createTree

Arguments

:: [HuffmanTree]
List of HuffmanTrees

-> HuffmanTree
Resulting HuffmanTree

The createTree function recursively creates one HuffmanTree from a list of HuffmanTrees. | It merges the two HuffmanTree with the lowest Weight, adds them to the list of HuffmanTrees and repeats until only one tree is left

encodeTree

Arguments

:: HuffmanTree
Completed HuffmanTree

-> [(Char, [Int])]
List of tuples containing character and path (binary code)

encodeString

Arguments

:: String
Text

-> Map Char [Int]
Map (AKA dictionary) where key is a unique character, and value a binary code

-> [Int]
Huffman encoded Text

The encodeString function completes Huffman Encoding on text using a Map (see encodeTree and subsequently tupleToMap)

decodeList

Arguments

:: [Int]
xss. Remainder of the text after head is removed for xs.

-> Map [Int] Char
Map (AKA dictionary) where key is a binary code, and value is a unique character

-> String
Huffman decoded Text

The decodeList function decodes a Huffman code using Huffman Codes in the form of a Map. | Checks if xs is a member of m. | If it is a member, it grabs that values calls itself again with an empty xs and xss as it is. | If it isn't a member of m, it calls itself where xs = xs and the first value of xss, and xss is the tail of xss

tupleToMap

Arguments

:: [(Char, [Int])]
List of tuples containing character and binary code

-> Map Char [Int]
Map (AKA dictionary) where key is unique character, and value binary code

The tupleToMap function is a wrapper for Data.Map.fromList. This is necessary for encoding the original text.

codedString :: String -> [Int]