Lib

Description

This module contains all necessary functions to encode to and decode from Huffman codes. This module was created for the Utrecht University of Applied Sciences AI course 'Declarative Programming'

Documentation

countValues

Arguments

:: String
The text to be encoded

-> [(Char, Int)]
(Unique character, frequency of said character)

The countValues function counts each unique character in a string. Adapted from https://stackoverflow.com/a/3711137.

createNodes

Arguments

:: [(Char, Int)]
Unique character is text, frequency of character in text

-> [HuffmanTree]
List of Node

The createNodes function creates a Node for each tuple. The frequency of a character equals its Weight.

createTree

Arguments

:: [HuffmanTree]
List of HuffmanTrees

-> HuffmanTree
Resulting HuffmanTree

The createTree function recursively creates one HuffmanTree from a list of HuffmanTrees. | It merges the two HuffmanTree with the lowest Weight, adds them to the list of HuffmanTrees and repeats until only one tree is left

encodeTree

Arguments

:: HuffmanTree
Completed HuffmanTree

-> [(Char, [Binary])]
List of tuples containing character and path (binary code)

The encodeTree function traverses (Depth First) a HuffmanTree. It then creates a code for each Node (or Leaf) Each time it goes left, it adds O to the path, every time it goes right it adds I to the path. When it reaches a Node, it returns a tuple containing the character and the path (the path needs to be reversed to be traversable in decoding).

encodeString

Arguments

:: String
Text

-> Map Char [Binary]
Map (AKA dictionary) where key is a unique character, and value a binary code

-> [Binary]
Huffman encoded Text

The encodeString function completes Huffman Encoding on text using a Map (see encodeTree and subsequently tupleToMap)

decodeList

Arguments

:: [Binary]
Binary code

-> Map [Binary] Char
Map (AKA dictionary) where key is a binary code, and value is a unique character

-> String
Huffman decoded Text

The decodeList function decodes a Huffman code using Huffman Codes in the form of a Map. | Checks if xs is a member of m. | If it is a member, it grabs that values calls itself again with an empty xs and xss as it is. | If it isn't a member of m, it calls itself where xs = xs and the first value of xss, and xss is the tail of xss

tupleToMap

Arguments

:: [(Char, [Binary])]
List of tuples containing character and binary code

-> Map Char [Binary]
Map (AKA dictionary) where key is unique character, and value binary code

The tupleToMap function is a wrapper for Data.Map.fromList. This is necessary for encoding the original text.

codedString

Arguments

:: String
String for encoding

-> [Binary]
Binary code representing string

Returns a Haskell encoded String Composition of encodeString, tupleToMap, encodeTree, createTree, createNodes and countValues

data Binary

Constructors

I

O